+++ templates/base.html.twig
@@ -157,7 +157,7 @@
 
                         <li class="list-inline-item copyright">
                             <span class="footer-below">
-                                copyright&copy; {{ "now"|date("Y") }}
+                                Copyright&copy; {{ "now"|date("Y") }}
                             </span>
                         </li>
                         <div class="col">
