+++ templates/main_controler/index.html.twig
@@ -49,7 +49,9 @@
                                         href="#deleteThisTrick">
                                         <i class="bi bi-trash-fill iconTricks"></i>
                                     </a>
-                                    <a href="{{ app.request.getSchemeAndHttpHost() }}/tricks/details/{{ trick.slug }}"
+                           
+                                    {# <a href="{{ app.request.getSchemeAndHttpHost() }}/tricks/details/{{ trick.slug }}" #}
+                                 <a href="{{ app.request.getSchemeAndHttpHost() }}/tricks/details/modifications/{{ trick.slug }}"}
                                         class="float-right ml-1">
                                         <i class="bi bi-pencil-fill iconTricks"><small><i class="bi bi-arrow-right-short"></i></small></i>
                                     </a>
