+++ templates/tricks/add.html.twig
@@ -28,7 +28,7 @@
 
                 <div class="form-group">
                     {{ form_row(formAddTrick.picture, {
-                        label: 'Add a picture',
+                        label: 'Add a main picture',
                         type : "file",
                         attr : {'class': 'form-control'},
                         }) }}
